Updated detail html with boostrap, so I don't need the detail css no more.
I have linked the home page to the detail page so when you click on new details it should take you to that detail html page.

Lastly on home I have added an image, so we can have that effect of the footer going up and the image background behind it. there is a white gap between the cards and the new arrivals I need to see how to fix that.